Improve tasks related to generating the app key
Modifies various tasks and templates to improve the efficiency and
indempotency of tasks related to setting the Snipe-IT app key. The app
key is now stored and read from a file to keep from modifying the
templated .env file. This removes the need for complex logic and reduces
errors, and improves indempotency.