Refactor role for indempotency and security
Overhaul's the role to make it more indempotent and to improve the
Snipe-IT upgrade process. Upgrades should now work simply by adjusting
the version variable, which changes the git tag checked out. Also
changes a number of permissions to increase security and improve
indempotency. Additional changes for clarity and repeatablility have
also been included.

The role currently is not entirely indempotent due to the existence of
CRLF files in the Snipe-IT GitHub repositiory. This causes an error if
the 'force' parameter is not used when cloning the repo, which reduces
indempotency. Alternative solutions would be to fork the repositiory or
change the CRLF files to LF.